New 2D plot style "with hsteps" (backport from 6.1)
The plot style "with hsteps" produces a step-like plot with
N horizontal line segments corresponding to N sequential data points.

Syntax:
	plot <data> with hsteps {forward|backward}
		{baseline|pillar|link|nolink}
		{{above|below} y=<baseline>}
		{offset <offset>}

The drawing elements of this style are

(1) Line segments extending horizontally from a data point to one or both sides

(2) Connecting lines linking two adjacent horizontal line segments

The hstep style draws a curve consisting of N horizontal segments and their
connecting lines.  The connecting line is a decoration that can take several
forms depending on the option {baseline|pillar|link|nolink}.
No symbol is placed on the data point itself.

A baseline y value is associated with each plot.  This baseline defines
the extent of vertical lines for options "baseline" or "pillars".
The area between the horizontal step and the baseline can be filled
using the standard gnuplot fill style mechanism.

The entire plot including the baseline maybe shifted along the y axis using
the "offset" keyword.  This allows stacking plots vertically for multiple
data sets that use the same internal y coordinate range.

A vital feature of the hsteps style is the ability to control the direction
of data points and horizontal segments drawn from them.
Similar styles already exist, such as steps, fillsteps, fsteps, histeps,
and boxes, but this new style is not intended to replace them.
Instead, it was created to compensate for the representation of graphs that
these styles cannot cover (see demos and documentation figures).
As a further feature, it gives a precise meaning to empty lines, NaN values,
and missing data in the input data so that they are evident in the drawing.

The new code is used to emulate existing step shape styles
(steps,fsteps,histeps,fillsteps).  The previous code for these styles is
left intact but deselected by #if (0) ... #endif.