backport reimplementation of local variables from 6.1
 - Add a locality flag to udvt_entry
 - When a local variable is declared, locality is set to current depth
 - For a global variable this should always be 0
 - Local variables are always added at the head of the variable list,
   which means during a lookup they are encountered before a global,
   or less-nested local, variable with the same name
 - When local depth is exited, step through variable list and delete all
   entries whose locality matches the depth

 Reverts or replaces previous implementation in commits
    a4e913218b	New feature: local variables
    89a9ddc896	local arrays need special treatment
    906efe72f1	Track names of dummy parameters used in defining a functionblock

This implementation is faster than the original for several reasons.
- local variables are kept at the front of the user variable list
  so references inside that scope will find them first
- variables ARG0 .. ARG9 are not touched by function block evaluation,
  so there is no need to restore them on exit from the function block.
  This saves work in misc.c:lf_pop()

The scope of a local variable is now confined to the file or
function block in which it was declared, including any bracketed
clauses within that same file or block.

Example:
    function $f1() << EOF
    local TEMP = 42
    return $f2()
    EOF

    function $f2() << EOF
    show var TEMP  # will list all matching global and local variables
    return TEMP
    EOF

    TEMP = 0
    answer = $f1()
    print "$f1() returns ", answer

./gnuplot foo.gp
        Variables beginning with TEMP:
        TEMP = 42       (locality 2)
        TEMP = 0
$f1() returns  0

The local definition and current value of TEMP inside $f1 is not
used by $f2 even though $f2 was called from inside $f1.

This works because each declared variable is tagged with the frame depth
("locality") of the LFS structure that was active when it was declared.
When a variable is referenced by name, a local variable with that
name is only preferred over a global variable with the same name
if the locality of the current LFS matches the locality of the variable.
Every call/load/functionblock invocation changes the locality.
However bracketed clauses have the same locality as the file or block
they appear in.