backport support for iteration over array content from 6.1
The iterator looks like this:
	array A = [1, 2, 4, 8]
	do for [a in A] { print a }

Thus it looks just like an iteration over words in a string,
but instead it uses successive entries in an array.

No checks are made for the type of the array entries.
You will, however, get the usual error messages if you try to apply
arithmetic operations to a string value or vice versa inside the
iteration loop.

This also fixes a parsing bug that meant you could not initialize a
matrix with an element of another matrix, e.g.

   array B = [1, 2, A[3], 4]