wxt:  remove --with-wx-multithreaded configuration option
back-ported from 6.1

The wxt terminal code has kept two code paths; one that claims to be
single-threaded and one that is explicitly multi-threaded.
The multi-thread code has not been working under linux for quite a while.
Various bug threads and ChangeLog entries blame it either on a change in
software architecture (wxgtk2->wxgtk3) or mutex failure on newer hardware
(Haswell CPUs).  Whatever the cause, it results in a crash or deadlock when
the wxt terminal is used.  This commit removes the broken code path.

So far as I can see this also means that the special sigint handling
routines are no longer needed.  These caught and deferred ctrl-C handling
inside many small sections of the wxt gui code.

That leaves the issue of handling software exceptions when int_error()
returns to top-level processing via a LONGMP in bail-to_command_line().
This is now handled on linux+gtk via a separate exception handler
(commit 255088ee).

These changes should not affect the Windows version of the wxt terminal,
but this remains to be tested in practice.