Do not save extraneous commands to $GPVAL_LAST_MULTIPLOT
Version 6 saves the commands that create a multiplot in a datablock
$GPVAL_LAST_MULTIPLOT. However if the commands are embedded in a bracketed
clause then the entire bracketed clause is saved rather than just the
multiplot commands. This may be harmless or it may cause a non-recoverable
error depending on what other commands are in that clause.

This commit searches the clause for the initial "set multiplot" and
the terminating "unset multiplot" and saves only that section of the
clause into $GPVAL_LAST_MULTIPLOT.

Bug #2714