Remove VMS-specific code
Gnuplot historically supported many platforms, both at the hardware
level and the O/S level, that no longer exist outside of museums or
historical hobbyist settings.  Terminal drivers and core support for
amiga and apollo systems was removed in 2010 (gnuplot version 4.4).
Support for sunview/Solaris in 2016 (gnuplot 5.2).  RIP also atari,
hercules, raw linux console mode, OpenStep, and 16-bit DOS.
In these cases it was reasonably clear that the hardware no longer
existed, the O/S running on it was no longer available, or in the
case of 16-bit DOS gnuplot's code was no longer compatible.

The end-of-life timeline for VMS has been less clear.
Even after the VAX and Alpha hardware it originally ran on went out
of production, DEC went out of business, and the inheritor of
commercial VMS (Hewlett Packard) dropped it also, VMS lived on as
OpenVMS through a port to run on x86 either directly on via an emulator.
But can we realistically claim that gnuplot still supports it?
I stopped being able to test build and debug gnuplot under VMS when the
last machine I had running VMS-on-alpha was decommissioned in about 2010.
The last successful build reported to the gnuplot mailing lists was
in December 2013 (gnuplot 4.6).  As of 2024, VMS Software Inc still
licenses and supports VMS on x86_64, but the only gnuplot version found
on their web site is a binary version of gnuplot 5.0 patchlevel 2
built in 2017.  A lot has changed in the gnuplot codebase since then.

Would the version 6 source even compile under OpenVMS 9.2?
Does anyone care?  Unless someone steps up to champion continued update
and support of the VMS-specific code, I think it is only realistic to
drop it from the gnuplot development version.

This commit removes all VMS-specific code and removes all claims that
this version of gnuplot runs on VMS.  The code remains in the stable
branch for gnuplot 6.0.

 31 files changed, 37 insertions(+), 2581 deletions(-)