protect cexp(z) from underflow
cexp(z) underflows when creal(z) < E_MINEXP
This was previously being caught as a floating point exception by
fetestexcept(FE_UNDERFLOW),  but apparently something has changed in
glibc or the c compiler such that cexp now sometimes sets errno=ERANGE
and returns NaN instead of zero and does not set a floating point
exception.  This is annoying.

The change was easily visible in the 2nd plot of igamma.dem
where the upper surface disappeared (NaN) rather than reaching
a plateau value of 1.

So now we test for creal(z) < E_MINEXP as a special case
and reset both errno and the return value cexp(z) to zero.
This is implements as a wrapper routine gp_cexp(z).
The new routine replaces cexp() in Igamma and LambertW.
Other call sites for cexp() remain, but I am not certain that they
are prone to underflow and, if they are, what is the appropriate
treatment.