"plot ... if (<expression>)"
Move the check for an if condition out of table-specific code into
the main parsing loop for "plot" commands.
The "if" restriction is accepted only for data plots.

Input lines that satisfy the if condition are processed as usual.
Lines that fail are essentially ignored, equivalent to matching
the "set datafile missing" condition".

TODO:  It might be useful in some cases to have the rejects treated as NaN/
       undefined rather than missing.  Perhaps this could be indicated by
       the filter expression returning NaN rather than 0 or 1.
       It's not obvious how to do this while preserving the actual input
       content however.