3D polygon objects can have individual border color and linewidth
The limitation that all 3D polygons share a single set of border
properties taken from "set pm3d" has been documented as an implementation
glitch to be removed at some point.  Now it's gone for polygon objects.
The border linewidth and color for individual 3D polygons is now taken
from the the values given in "set object".

The polygons demo is changed accordingly to set the border linewidth
and color via "set object".

The documented limitation still holds for "splot with polygons".