Rework term->waitforinput() to accommodate UTF-8 with BSD editline
This is intended to fix the problem that when gnuplot was built
    ./configure --with-readline=bsd
it became impossible to enter UTF-8 character sequences from the terminal.
It still worked to read or write UTF-8 encoded files.
This is due to incomplete functionality in the shim code and header
<editline/readline.h> provided with editline that provides an
almost-good-enough equivalent for the gnu libreadline API.

The readline shim header for the BSD editline library marks as
"not implemented" the API element
	int (*rl_getc_function)(FILE *)
that can be used to interpose a user-supplied function between the library
and a call to getchar().  While it does in fact provide a hook by that
name for a wrapper function, the hook does not substitute directly for the
libreadline equivalent because the editline library expects wide character
input rather than a multi-byte UTF-8 sequence.

We can work around this by having gnuplot's term->waitforinput() routines
call back into the guts of the editline library to feed it a wchar built
from the UTF-8 byte sequence.  That requires mixing calls to the readline
shim with direct calls to el_wgetc().  So far as I can see this mixture
is undocumented, but it does seem to work.

Tested with libedit 3.1 20221030 and 20230828

plot.c
	initialize editline with both el_init() and history_init()
readline.c
	new macro read_and_return_character() invokes either
	getchar() or el_wgetc() as appropriate.
term.c
x11.trm
caca.trm
wxt_gui.cpp
qt_term.cpp
	use the new macro in term->waitforinput()