Do not draw variable color points with color value NaN
Gnuplot version 5 marked a point as invalid (coordinate type UNDEFINED)
on input if any of the columns in a using specifier held NaN.

Gnuplot version 6 does not do this if the column is optional, as in the
case of variable color.  Instead it leaves it to the plotstyle-specific
code to decide what to do with a NaN or otherwise out-of-range optional
value.  For example the image code treats a pixel with value NaN as
fully transparent, and thus does not draw it.  Most (all?) styles set
the line type to NODRAW, but this doesn't affect points.

The point drawing code should do the same as the image code -
do not draw a point if its color value is NaN.

Bug 2737