improved multiplot mousing (1/4) - limit mouse events to active panel
Only the final (last-drawn) panel of a multiplot is active for mouse
button and wheel (pan/zoom) operations.
Ignore mouse button and wheel events from inactive regions.
When the active panel is panned or zoomed, do not use the new axis
range setting to override pre-set axis ranges in the inactive panels.

However if the plot in an inactive panel does not specify axis ranges
separately then it may be affected by the zoom operation because these
values are still in effect at the start of the multiplot replay operation.
It may be possible to further improve this behavior, perhaps by filtering
for incompatible settings like log-scale.

Part of this change replaces
	TBOOLEAN inside_zoom
with a function
	TBOOLEAN inside_zoom(void)
In the absence of mouse support this reduces to constant FALSE.