improved multiplot mousing (2/4) - save/restore shadow axis structures
During multiplot zoom, save the original shadow axis state in parallel with
the primary axis state.  This was previously noted as a possible alternative
but it has become clear that it is necessary so that the "via" and "inv"
function pointers are not left dangling when the axes are reinitialized.

When the multiplot is un-zoomed (hot key "u" or successive "p"), the original
state is restored.