prevent possible use-after-free if a function block has no "return" statement
When a function block is called by the evaluation stack processing routine
f_eval() it retrieves the return value from a temporary structure and pushes
it onto the stack.  That return value was supposed to have been loaded by
a "return <value>" statement in the function block.  If, however, the
function block contains no "return" statement it is possible that the
content of the temporary structure is left over from a function block
call deeper in the evaluation stack.   This can lead to a use-after-free
error when the content is "pushed" a second time.

Now we reset the temporary structure value to UNDEFINED immediately after
pushing it onto the stack.  Even if it is erroneously referenced again
afterwards there will be no use-after-free.

Bug 2745