watchpoints: rework implementation of watchpoint functions
Previously the only legal form was
	plot foo watch f(x,y)=target
where "f(x,y)" was a required string and f was a pre-defined function.

Now any function is legal, and parameters can be arbitrary.
	plot foo watch func(a,b,c,d)=target

Function evaluation uses the current values of x, y, and z regardless
of whether they were explicit paramters, so it is possible to do
	contourlevel(foo)=z
	plot foo watch contourlevel("dummy")=60.