Filter "bins" with "binvalue=avg" should report NaN rather than 0 if the bin is empty
This modifies the handling of bins that contain no data when 'binvalue=avg'.
In the current implementation, missing values default to 0.
This changes that behavior by setting them to NaN instead.
Internally, the point type for such bins is also set to UNDEFINED.
When 'binvalue=sum', the value of bins that contain no data are left as 0.