display key sample in "with pm3d" style with 2-sided fillcolor
The style "with pm3d fillcolor linestyle N" produces a 2-sided surface
with separate colors for the top and bottom.  If the key show a single
color, at some view angles the entire visible surface might be a mismatch
with the key sample.

With this commit, for the special case of "fillcolor ls N"
the top and bottom colors are represented in the key sample by the upper
and lower halves of a rectangle divided along the diagonal.