Fix incorrect handling of variable fill color for 3D boxes
In 3D plots variable color is stored on input in an aliased field of
the coordinate structure as a (double) value.
However for 3D boxes in particular it is passed to the rendering code
by temporarily copying it to plot->lp_properties.pm3d_color.lt,
which is a signed integer.
This caused an error in rendering if the color is a 32 ARGB value with8
the high bit set (alpha > 127).

Casting the color value to (unsigned) while copying fixes this in practice,
A better fix would provide a separated handling path for color based on
linetype (signed value) and color provided as ARGB (unsigned value).

Bug 2812