cast colorspec.lt to (unsigned int) when it may hold 32-bit ARGB
Bug reports 2812 and 2813 diagnose a problem with color handling that
manifests on ARM architectures but not on x86.

The basic issue arises when a 32-bit ARGB color with the leading bit set
(opacity value < 0.5) is stored as a (double) data value and then
later this value is retrieved for use as a 32-bit color.
If the 32-bit value was originally held in a signed integer variable,
it looked like a negative number and thus was stored as a negative
floating point value in the (double).  If this is later assigned to
an unsigned int variable, ARM truncates the value to zero.

Conversely if the 32-bit value was originally held as an unsigned
integer and thus stored as a positive value. On later assignment
to a signed integer variable there was overflow and ARM may truncate
the value to MAXINT.

To avoid this we must explicitly cast the original ARGB value to
(unsigned int) so that it is a positive number before storing it in
a double, and take care also to retrieve it as an unsigned int so that
there is no overflow.