Support for cross-compilation for Windows on ARM
The new Makefile in the config/woa directory is a modified copy of the
mingw Makefile with changes for cross-compilation using the
MSYS2/CLANG64/CLANGARM64 environment.

The build includes all options which are included with the MSYS2/CLANG64
build for x64 Windows. It is hence currently built without linking to
openspecfun as this requires fortran.

The copydlls.rex file is a REXX script to recursively call 'objdump' to
locate all required DLLs for distribution. The approach on x86/x64 using
'ldd'  unfortunatly does not work here. The script should eventually be
converted to a more widely available script language.