resolve ambiguous syntax in "set dashtype"
Statements of the form
    set for [i=1:N] dashtype i (n, m)
fail because it is parsed as if it invoked a function i(n,m).
Resolve this by checking first whether i is a known variable,
and if so dereference it directly without parsing it as an expression.

This is not a real fix, just a patch over bad language design.
An expression like
    set dashtype i+j (n,m)
will still fail

Bug 2814