Support use of either ZEXINT or CEXINT (Amos library expint routines)
The Amos library routines for complex special functions were originally
written in two forms, a z***.f form that processed the complex values
using separate real and imaginary variables (Fortran DOUBLE PRECISION),
and a c***.f form that processed the values as a single Fortran COMPLEX.
This was because at the time they were written, Fortran did not support
double precision complex variables.  Now it does.

The Amos implementations of the complex exponential integral function
(CEXINT and ZEXINT) are not in libopenspecfun, which is the most easily
available support library for this class of functions on linux.
Thus gnuplot support for complex expint() is present only if one
of these routines is found by the configuration script, leaving it up
to the user to provide a library that contains it.
Up until now the gnuplot code assumed that a double precision CEXINT
would be the routine provided; now either CEXINT or ZEXINT is accepted.

ZEXINT support requested by and initial version provided by Erik Luijten