Add code in the sharpen filter to handle step functions
The sharpen filter is intended to detect cases where a sharp peak
or discontinuity in a plotted function is missed or distorted
because the sampling is too coarse to catch it.
The approach used to detect such features is based on approximating
the second derivative of the function. The local extremum is then
found by bisection. This approach fails to handle step functions
because bisection to find an extremum is meaningless in this case.

This commit adds a separate test for the vertical jumps in a
step function. Instead of bisection to find an extremum, the
program now bisects these to find the x coordinate of the jump.