revisit escape sequences (octal)
The documentation has [always?] stated that octal escape sequences
are of the form \ooo where 'o' is a valid octal digit.
However that is not what the code has actually done.  It accepted from
one to four digits.  Thus for example \7 and \0106 were accepted even
though neither conforms to the stated restriction. \456 was accepted
even though values above \377 cannot fit in a one byte character.

Accepting one- and two- digit sequences leads to strangeness like this:

    gnuplot> set title "\A\B\1\2\3\4\5\6\7\8\9\0\h\i\j\k\x\y\z"
    gnuplot> show title
             title is "AB89"

The stored string actually contains AB^A^B^C^D^E^F^G89 but the
middle 7 characters either don't print at all or trigger bizarre
output depending on the terminal.

Furthermore the documentation claims that escape sequences are not
expanded in single-quoted strings.  But this was [never?] true for
enhanced text strings.

I think the original rationale for octal escape sequences was to
provide some way to specify characters from the Adobe Symbol font
by code point, since it doesn't use a normal encoding.
In the era of UTF-8 encoding and non-Adobe symbol fonts this
rationale is outdated. It can even be counterproductive because the
documentation leads users to believe that "{/Symbol \245}" will
produce ∞ (infinity), but this is only true for the Adobe Symbol
font while most systems now provide a different font also named
"Symbol" but using a standard encoding and a different set of
symbols.

Truly fixing all this is probably impossible, but also no longer
necessary since using Unicode code points instead works more
consistently.

Nevertheless I think it is worth cleaning up the escape sequence
processing to the extant that only 3-digit sequences from 000-377
are interpreted as octal codepoints.  This at least means that
single characters escapes are consistent, e.g. "\7" yields the
character "7" rather than the control code ^G (BEL), and "\0"
yields the character "0" rather then terminating the string or
not depending on whether enhanced text is active.

There remains a difference between the documentation and the code
as to what happens to escape sequences in single-quoted enhanced
text.