Do not disable mousing after a ^C
History:
- The restriction to disable mousing was present in version 4,
- disabled in versions 5.0 and 5.2 with a comment that any terminal
  with potential mouse-before-initialized issues should check status for itself,
- reinstated in 5.4 without comment. Checks for terminal initialization
  were added in other places at about the same time, so this may have been
  an over-cautious inclusion.

The loss of mousing after ^C is annoying.
This commit disables this restriction again (as in versions 5.0 - 5.2).
The intent is that if a problem is reported for a particular terminal
the fix should be made terminal-specific.

Note:
The underlying issue is that the ^C processing (SIGINT) in term_reset()
sets "term_initialised = FALSE".  This flag is also set to FALSE
in set_output(), but I think terminals that direct output via
"set output" do not track mousing anyway.

It does make sense to not attempt mouse tracking on an uninitialised terminal,
but that's not really the case for an interactive terminal after a ^C.
So perhaps a better fix would be to modify the SIGINT processing to not
call term_reset()?  But that seems more likely to cause problems than the
change here.