do not break contour lines into 100-segment fragments
Problem:
The 3D contour generation algorithm has imposed a limit of 100 contiguous
line segments within a contour.  If the contour continues, it starts a
new fragment and resets the segment count to zero.
This is fine for the contour lines themselves, but it causes problems if
the contour line is saved for later use as a polygon boundary.
Example
	set contour surface
	set table $CONTOURS
	splot f(x,y) with lines nosurface
	unset table
	unset contour
	splot $CONTOURS with polygons fs solid border

Any contour line that was broken into multiple fragments because of the
100-segment limit would be rendered as multiple polygons whose borders
include artifactual lines connecting the two endpoints of each fragment
to an inappropriate vertex rather than merging adjacent polygons.

Fix:
Dynamically allocate the list of points making up a contour line.
Allow it to grow as needed until the contour either becomes a closed
path or terminates at the plot boundaries.

Possible downside:
Some terminals may handle very long (thousands of segments?) contours
poorly, since accumulated error of short relative vectors may lead to
a gap at the point of closure.

Feature Request 591