fix bug in calculation of a**b for integer a, integer b < 0
Since the result may not be expressable as an integer the code
in f_power made a floating point calculation instead, but incorrectly
treated this as an overflow.  For the default overflow handling
this was not a problem.
However for "set overflow NaN" or "set overflow undefined"
the result was incorrectly set to NaN or undefined.
E.g.
    set overflow NaN
    print (-2)**(-2)
	NaN

Correct result
    print (-2)**(-2)
	0.25