Revise the definition and handling of a plugin's destructor function
To recap the way a plugin works:

An imported function (a.k.a. plugin) can optionally provide an
initializer function and a destructor function. Until now gnuplot has
only checked for and invoked the initializer function.
Now the corresponding destructor is also handled.

The initializer is found by requesting a function named gnuplot_init
from the same shared object as the imported function.
The destructor if found by requesting a function named gnuplot_fini.
For example, the command
      gnuplot> import newdata() from "demo_plugin"
searches the shared object demo_plugin.so (or demo_plugin.dll) for
a function named newdata.  It also looks for functions gnuplot_init
and gnuplot_fini, storing them as part of the newdata definition.
The initializer, if any, is invoked immediately.

The initializer can allocate and initialize any data structure, files,
pipes, etc, that will be needed when the plugin function is later called
from gnuplot.  These would typically be wrapped into a single dynamically
allocated structure whose address is passed back to gnuplot as an opaque
token.

When the imported function is called from gnuplot, the opaque token is
passed together with the user-visible function parameters.

The destructor is called by gnuplot when the plugin function goes out of use.
This happens if
- the function is redefined via another "import"
- the function is redefined from the command line
- the program exits normally (FIXME: not yet implemented)

Both the initializer and the destructor are also passed the address of
the imported function they are associated with.  This allows a single
gnuplot_init() to service multiple or alternative imported functions
from the same shared object.

See template function types at the head of external.c

FIXME:
- The destructor should also be called on program exit.
- This all needs proper documentation. As this point the only guide
  is the source code in the demo/plugin directory.