revised color assignment code for "splot with polygons"
[modified back-port from 6.1]

The fill color and border properties of "splot with polygons"
were often incorrect, or at least unexpected.
Equivalent plots made by the development version are more in line
with expectations.  However the code in 6.0 and 6.1 has diverged
enough that a complete back-port would be non-trivial.
This set of changes brings the fill color and border color into
agreement, but the inheritence of other line properties
(default color, line width) is still unreliable.