Warn about libgd font issues in the 6.0.4 Release Notes
I have finally pinned down the source of recurring problems with
rotated text, particularly vertical text, in the libgd terminals.
It is due to libgd's use of the libraqm shared library to determine
glyph spacing. Several versions of these two libraries did not play
nicely together.  However in the latest versions (built today from
the respective git HEADs) the problems have been resolved.
These working versions report as
	libgd   version 2.4.0
	libraqm version 0.10.3

I do not know whether these reported versions are sufficient to
identify "good" versions in various linux repositories.

It is too late to do anything about this in the distribution
packages for gnuplot 6.0.4 but I have updated the Known Issues
section of the on-line release notes to include this information.