gd: Revise font handling to match libgd 2.4.0 and libraqm 0.10.3
Positioning and kerning of rotated text, in particular vertical text,
has been buggy in the upstream gd and raqm libraries for several
versions. The version of libharfbuzz may also be relevant since
libraqm uses it.

Things started working again [mostly] in libgd 2.4.0 used with
libraqm 0.10.3, but the information reported for vertical
text extent is still off.  This commit changes the put_text() code
to not use the reported vertical text extent.  With this in place
and with the updated libraries, everything seems to work again.

Some earlier version pairings (libgd 2.3.x libraqm 0.10.x) may also
work, but no guarantees.