multiplot mousing:  Finally! mouse readout for all panels of a multiplot
The resolves one of the two oldest and most frequent user requests.
Since the multiplot option was first introduced into gnuplot
(in version 3.7) it has been a recurring feature request from users
to support mouse interaction with all panels in the multiplot.
Neither mousing nor "replot" were possible, however, because the
context and data structures used for each panel were lost when the
plot command for the next panel was issued.

Version 6.0 reworked the program to save the command sequence that
created the multiplot into a data block so that it could be replayed
later by a "replot" command or during mousing.  It remained a limitation
that only the most recently drawn panel was active for mouse interaction
because the necessary information, in particular the axis settings,
were only saved for the final plot.

This commit finally allows mouse tracking for all panels in a multiplot.
Now we keep an array of panel_bounds and parallel arrays of axis
settings for each panel.  As the multiplot is drawn, the relevant
bounding box and axis settings for that panel are saved.
During mousing, the current mouse coordinates are matched against the
array of panel bounds.  If there is a match, the mouse coordinates
are converted to user coordinates using the axis mappings saved for
that panel.

TODO
- The data arrays have fixed dimension MAX_PANELS (currently 16).
  They could instead be dynamically allocated under user control.