multiplot mousing:  classify panel content in panel_flags[]
The global flags is_3d_plot and splot_map are not valid for
a multiplot since they refer to the most recent "splot" command
rather than to the content of any particular multiplot panel.

Replace these with flags PANEL_3d PANEL_SPLOT PANEL_2D held
in a per-panel array panel_flags[MAX_PANELS].

Replace the tests for is_3d_plot and splot_map with routines
that are aware of multiplot panels:
  is_3d_plot_or_panel()
  splot_map_or_panel()

Replace all instances of the macro ALMOST2D in mouse.c with
some combination of the above.  This leaves no users of the
macro in the core code, only in individual terminals.