save:  distinguish between blank axis label and no axis label at all
On entry, or after "unset x2label", the save commands now emit
	unset x2label
rather than
	set x2label ""

This is also true for the other axis labels and for title and timestamp.
Mostly this does not make a difference if the save file is read back in,
but in the case of x2label and y2label, the layout for "splot map"
because space is erroneously reserved for the x2/y2 labels and hence the
size left for the plot itself is reduced.