save: If the visible axis properties have not changed, save as "unset"
Not all axis properties are visible, and only the visible properties
are saved.  That means that loading a file that contains
    set xrange [*:*] noreverse
which is what the "save" command produces, does not update any of
the hidden properties min, max, data_min, data_max, ...
So the result of loading depends on the current program state.
That is undesireable if you are trying to reproduce a previously
saved figure.

Now we fix this by checking whether the visible axis properties
match those at program entry, and if so save "unset xrange" instead.