multiplot mousing:  queue pan/zoom event from hotkey and apply it later
Rather than trying to apply a pan/zoom request immediately when it
is triggered by a mouse event, note the panel in which the event
ocurred and queue the operation for the next time the panel is redrawn.
Normally the event also causes a multiplot playback sequence that
will redraw it.
- the initial program state is restored by loading $GPVAL_PRE_MULTIPLOT
- the initial panels are drawn as before
- just before drawing the target panel the queued zoom is applied
- the target panel is drawn
- the multiplot completes, leaving the zoomed state in place

TODO:
  motion events in 3D plots should either be treated similarly